+++ src/p_001.rs
@@ -1,4 +1,4 @@
 #[allow(dead_code)]
-pub fn problem_001() -> i128 {
+pub fn problem_001() -> i64 {
     (0..1000).filter(|x| x % 3 == 0 || x % 5 == 0).sum()
 }
