+++ src/p_007.rs
@@ -1,5 +1,5 @@
-fn is_prime(n: i128) -> bool {
-    let n_int_sqrt = (n as f64).sqrt() as i128;
+fn is_prime(n: i64) -> bool {
+    let n_int_sqrt = (n as f64).sqrt() as i64;
 
     for i in (3..=n_int_sqrt).step_by(2) {
         if n % i == 0 {
@@ -10,7 +10,7 @@ fn is_prime(n: i128) -> bool {
 }
 
 #[allow(dead_code)]
-pub fn problem_007() -> i128 {
+pub fn problem_007() -> i64 {
     let mut prime_count = 4;
     let mut n = 11;
