+++ src/p_014.rs
@@ -1,11 +1,11 @@
 #[allow(dead_code)]
-pub fn problem_014() -> i128 {
-    let mut max_chain_size: i128 = 0;
-    let mut max_n: i128 = 0;
+pub fn problem_014() -> i64 {
+    let mut max_chain_size: i64 = 0;
+    let mut max_n: i64 = 0;
 
     for n in 3..1_000_000 {
-        let mut chain_size: i128 = 1;
-        let mut last_chain_n: i128 = n;
+        let mut chain_size: i64 = 1;
+        let mut last_chain_n: i64 = n;
 
         while last_chain_n != 1 {
             if last_chain_n % 2 == 0 {
