+++ src/p_016.rs
@@ -1,4 +1,4 @@
-fn explode_carry(mut digits: Vec<i128>) -> Vec<i128> {
+fn explode_carry(mut digits: Vec<i64>) -> Vec<i64> {
     let max_index = digits.len() - 1;
 
     for i in 0..digits.len() {
@@ -22,8 +22,8 @@ fn explode_carry(mut digits: Vec<i128>) -> Vec<i128> {
     digits
 }
 
-fn big_pow(base_n: i128, p: i128) -> Vec<i128> {
-    let mut digits: Vec<i128> = vec![1];
+fn big_pow(base_n: i64, p: i64) -> Vec<i64> {
+    let mut digits: Vec<i64> = vec![1];
 
     for _ in 0..p {
         for n in &mut digits {
@@ -37,6 +37,6 @@ fn big_pow(base_n: i128, p: i128) -> Vec<i128> {
 }
 
 #[allow(dead_code)]
-pub fn problem_016() -> i128 {
+pub fn problem_016() -> i64 {
     big_pow(2, 1000).iter().sum()
 }
